+++ CHANGELOG.md
@@ -7,10 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - Add application's support for Nextcloud 31
+
+## 2.7.1 - 2024-10-31
+### Changed
 - Make error handling better in `integration_setup.sh` file for integration configuration setup.
 - Improve UI by using Nextcloud's `NoteCard` in project folder setup error.
 - Resolve the issue with retrieving the Nextcloud server version for version compare
 - Add warning UI when encryption is not explicitly enabled for `groupfolders`.
+- Fix hash or encrypt secret for different nextcloud versions
 
 ## 2.7.0 - 2024-09-10
 ### Changed